Move the generated content to template
Since we generate go code, the template become a noise because its inline
with the generator itself, which make the code unreadable. This commit
split and move the content of generated to its own file `template.go`.

Also, replace Fprintf() with Fprint() on formatted string that does not
have argument.