cmd/go-bindata: add verbose option
The verbose option ("-verbose") print the file that being scanned.
For file that is being added (either with include option or not) it will
print the path prefixed with "+", and for file that is ignored it will
print the path prefixed with "-".

The last line will print the name of generated Go file.

Close #21