all: use map to collect list of assets instead of slice
In issue #36 and #37 we found that if one of the inputs is duplicate
the generated Go code will also contains duplicate functions which
cause the code cannot be compiled.

We try to fix this in 53287067 but it cause the issue #28 reappear again.

The problem is we did not check if an asset already collected, because
we use slice to merge all output from Scan().

Previous code collect asset by using the following process,

	for _, input := range inputs {
		inputAssets := Scan(input)
		assets = append(assets, inputAssets)
	}

Given the following inputs: "a" and "a/b", there is no check for "a/b"
is already exists in assets.

This changes replace slice of asset with map of asset, indexed by
name. So, after the Scan() completed we add to assets only if the name
is not exist,

	for _, input := rage inputs {
		inputAssets := Scan(input)
		for name, asset := range inputAssets {
			_, ok := assets[name]
			if !ok {
				assets[name] = asset
			}
		}
	}

This change is actual fix #36 and fix #37, and I hope resolve #38.